+++ blocksuite/affine/block-root/src/transformers/markdown.ts
@@ -16,7 +16,7 @@ import { createAssetsArchive, download, Unzip } from './utils.js';
 
 function getProvider() {
   const container = new Container();
-  const exts = SpecProvider.getInstance().getSpec('store').value;
+  const exts = SpecProvider._.getSpec('store').value;
   exts.forEach(ext => {
     ext.setup(container);
   });
