+++ blocksuite/affine/block-root/src/transformers/notion-html.ts
@@ -14,7 +14,7 @@ type ImportNotionZipOptions = {
 
 function getProvider() {
   const container = new Container();
-  const exts = SpecProvider.getInstance().getSpec('store').value;
+  const exts = SpecProvider._.getSpec('store').value;
   exts.forEach(ext => {
     ext.setup(container);
   });
