+++ blocksuite/affine/block-surface-ref/src/portal/note.ts
@@ -118,7 +118,7 @@ export class SurfaceRefNotePortal extends WithDisposable(ShadowlessElement) {
       query: this.query,
       readonly: true,
     });
-    const previewSpec = SpecProvider.getInstance().getSpec('preview:page');
+    const previewSpec = SpecProvider._.getSpec('preview:page');
     return new BlockStdScope({
       store: doc,
       extensions: previewSpec.value.slice(),
