+++ blocksuite/affine/block-surface-ref/src/surface-ref-block.ts
@@ -239,8 +239,7 @@ export class SurfaceRefBlockComponent extends BlockComponent<SurfaceRefBlockMode
 
   private _previewDoc: Store | null = null;
 
-  private readonly _previewSpec =
-    SpecProvider.getInstance().getSpec('preview:edgeless');
+  private readonly _previewSpec = SpecProvider._.getSpec('preview:edgeless');
 
   private _referencedModel: GfxModel | null = null;
