+++ blocksuite/blocks/package.json
@@ -47,28 +47,7 @@
     "@blocksuite/global": "workspace:*",
     "@blocksuite/icons": "^2.2.1",
     "@blocksuite/inline": "workspace:*",
-    "@blocksuite/store": "workspace:*",
-    "@floating-ui/dom": "^1.6.10",
-    "@lit/context": "^1.1.2",
-    "@preact/signals-core": "^1.8.0",
-    "@toeverything/theme": "^1.1.11",
-    "@vanilla-extract/css": "^1.17.0",
-    "date-fns": "^4.0.0",
-    "dompurify": "^3.2.4",
-    "fflate": "^0.8.2",
-    "file-type": "^20.0.0",
-    "fractional-indexing": "^3.2.0",
-    "html2canvas": "^1.4.1",
-    "katex": "^0.16.11",
-    "lit": "^3.2.0",
-    "lodash-es": "^4.17.21",
-    "lz-string": "^1.5.0",
-    "minimatch": "^10.0.1",
-    "nanoid": "^5.0.7",
-    "shiki": "^2.0.0",
-    "simple-xml-to-json": "^1.2.2",
-    "yjs": "^13.6.21",
-    "zod": "^3.23.8"
+    "@blocksuite/store": "workspace:*"
   },
   "exports": {
     ".": "./src/index.ts",
@@ -79,13 +58,9 @@
     "src",
     "dist",
     "!src/__tests__",
-    "!dist/__tests__",
-    "schemas.d.ts",
-    "schemas.js"
+    "!dist/__tests__"
   ],
   "devDependencies": {
-    "@types/katex": "^0.16.7",
-    "@types/lodash-es": "^4.17.12",
     "@vanilla-extract/vite-plugin": "^5.0.0",
     "vitest": "3.0.6"
   }
