--- blocksuite/blocks/schemas.d.ts
@@ -1,2 +1,0 @@
-// @ts-expect-error FIXME: ts error
-export * from './dist/schemas';
