--- blocksuite/blocks/schemas.js
@@ -1,3 +1,0 @@
-/// <reference types="./dist/schemas.d.ts" />
-// eslint-disable-next-line @typescript-eslint/no-restricted-imports
-export * from './dist/schemas.js';
