--- blocksuite/blocks/src/_common/adapters/extension.ts
@@ -1,22 +1,0 @@
-import {
-  AttachmentAdapterFactoryExtension,
-  HtmlAdapterFactoryExtension,
-  ImageAdapterFactoryExtension,
-  MarkdownAdapterFactoryExtension,
-  MixTextAdapterFactoryExtension,
-  NotionHtmlAdapterFactoryExtension,
-  NotionTextAdapterFactoryExtension,
-  PlainTextAdapterFactoryExtension,
-} from '@blocksuite/affine-shared/adapters';
-import type { ExtensionType } from '@blocksuite/store';
-
-export const AdapterFactoryExtensions: ExtensionType[] = [
-  AttachmentAdapterFactoryExtension,
-  ImageAdapterFactoryExtension,
-  MarkdownAdapterFactoryExtension,
-  PlainTextAdapterFactoryExtension,
-  HtmlAdapterFactoryExtension,
-  NotionTextAdapterFactoryExtension,
-  NotionHtmlAdapterFactoryExtension,
-  MixTextAdapterFactoryExtension,
-];
