--- blocksuite/blocks/src/_specs/index.ts
@@ -1,6 +1,0 @@
-export * from './common.js';
-export * from './preset/edgeless-specs.js';
-export * from './preset/mobile-patch.js';
-export * from './preset/page-specs.js';
-export * from './preset/preview-specs.js';
-export { SpecBuilder, SpecProvider } from '@blocksuite/affine-shared/utils';
