--- blocksuite/blocks/src/_specs/preset/edgeless-specs.ts
@@ -1,9 +1,0 @@
-import { EdgelessBuiltInSpecs } from '@blocksuite/affine-block-root';
-import type { ExtensionType } from '@blocksuite/store';
-
-import { EdgelessFirstPartyBlockSpecs } from '../common';
-
-export const EdgelessEditorBlockSpecs: ExtensionType[] = [
-  EdgelessBuiltInSpecs,
-  EdgelessFirstPartyBlockSpecs,
-].flat();
