--- blocksuite/blocks/src/_specs/preset/page-specs.ts
@@ -1,9 +1,0 @@
-import { PageRootBlockSpec } from '@blocksuite/affine-block-root';
-import type { ExtensionType } from '@blocksuite/store';
-
-import { PageFirstPartyBlockSpecs } from '../common.js';
-
-export const PageEditorBlockSpecs: ExtensionType[] = [
-  PageRootBlockSpec,
-  ...PageFirstPartyBlockSpecs,
-].flat();
