+++ blocksuite/blocks/src/extensions/index.ts
@@ -1,0 +1,3 @@
+export * from './common.js';
+export * from './editor-specs.js';
+export * from './preview-specs.js';
