+++ packages/frontend/core/src/components/blocksuite/block-suite-editor/specs/custom/spec-patchers.tsx
@@ -32,11 +32,15 @@ import { I18n } from '@affine/i18n';
 import { track } from '@affine/track';
 import {
   BlockServiceWatcher,
+  type BlockStdScope,
   BlockViewIdentifier,
+  ConfigIdentifier,
+  LifeCycleWatcher,
   type WidgetComponent,
 } from '@blocksuite/affine/block-std';
 import type {
   AffineReference,
+  CodeBlockConfig,
   DocMode,
   DocModeProvider,
   OpenDocConfig,
@@ -44,6 +48,7 @@ import type {
   PeekOptions,
   PeekViewService as BSPeekViewService,
   QuickSearchResult,
+  ReferenceNodeConfig,
   RootBlockConfig,
 } from '@blocksuite/affine/blocks';
 import {
@@ -52,20 +57,23 @@ import {
   DocModeExtension,
   EdgelessRootBlockComponent,
   EmbedLinkedDocBlockComponent,
+  FeatureFlagService,
   GenerateDocUrlExtension,
   insertLinkByQuickSearchCommand,
-  MobileSpecsPatches,
   NativeClipboardExtension,
   NoteConfigExtension,
   NotificationExtension,
   OpenDocExtension,
+  ParagraphBlockService,
   ParseDocUrlExtension,
   PeekViewExtension,
   QuickSearchExtension,
   ReferenceNodeConfigExtension,
+  ReferenceNodeConfigIdentifier,
   RootBlockConfigExtension,
   SidebarExtension,
 } from '@blocksuite/affine/blocks';
+import type { Container } from '@blocksuite/affine/global/di';
 import { Bound } from '@blocksuite/affine/global/utils';
 import {
   type BlockSnapshot,
@@ -602,6 +610,66 @@ export function patchForSharedPage() {
 }
 
 export function patchForMobile() {
+  class MobileSpecsPatches extends LifeCycleWatcher {
+    static override key = 'mobile-patches';
+
+    constructor(std: BlockStdScope) {
+      super(std);
+      const featureFlagService = std.get(FeatureFlagService);
+
+      featureFlagService.setFlag('enable_mobile_keyboard_toolbar', true);
+      featureFlagService.setFlag('enable_mobile_linked_doc_menu', true);
+    }
+
+    static override setup(di: Container) {
+      super.setup(di);
+
+      // Hide reference popup on mobile.
+      {
+        const prev = di.getFactory(ReferenceNodeConfigIdentifier);
+        di.override(ReferenceNodeConfigIdentifier, provider => {
+          return {
+            ...prev?.(provider),
+            hidePopup: true,
+          } satisfies ReferenceNodeConfig;
+        });
+      }
+
+      // Hide number lines for code block on mobile.
+      {
+        const codeConfigIdentifier = ConfigIdentifier('affine:code');
+        const prev = di.getFactory(codeConfigIdentifier);
+        di.override(codeConfigIdentifier, provider => {
+          return {
+            ...prev?.(provider),
+            showLineNumbers: false,
+          } satisfies CodeBlockConfig;
+        });
+      }
+    }
+
+    override mounted() {
+      // remove slash placeholder for mobile: `type / ...`
+      {
+        const paragraphService = this.std.get(ParagraphBlockService);
+        if (!paragraphService) return;
+
+        paragraphService.placeholderGenerator = model => {
+          const placeholders = {
+            text: '',
+            h1: 'Heading 1',
+            h2: 'Heading 2',
+            h3: 'Heading 3',
+            h4: 'Heading 4',
+            h5: 'Heading 5',
+            h6: 'Heading 6',
+            quote: '',
+          };
+          return placeholders[model.type];
+        };
+      }
+    }
+  }
   const extensions: ExtensionType[] = [
     {
       setup: di => {
